Year 2 Group Project on Android written in Java
My role in this Group Project was more focused on overall structure of
the project and making sure the group new what to do and what we needed
to work on. A leader-esque role. This can be seen from meetings.txt
which was done by myself. Programming wise I focused on the Flight,
Itineraries, and FlightDatabase side of the project.